refactor(auth): delete interfaces that do not fit into a polymorphic structure
